pyproject.toml

[tool.poetry]
name = "gpu-reliability"
version = "0.1.0"
description = "A test harness for cloud GPU provisioning."
authors = ["Pierce Freeman <pierce@freeman.vc>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
google-cloud-compute = "^1.5.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pytest = "^7.1.2"

[tool.poetry.scripts]
benchmark = "gpu_reliability.cli:benchmark"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
combine_as_imports = true
force_grid_wrap = 4
float_to_top = true
include_trailing_comma = true
known_first_party = "gpu_reliability"
extra_standard_library = "pkg_resources"
skip = "__init__.py"

line_length = 120
lines_after_imports = 2
multi_line_output = 3
